requirements.txt
requests>=2.31.0
flask>=3.0.0
flask-login>=0.6.3
flask-wtf>=1.2.1
pyjwt>=2.8.0
